Add "elk list" command
