+++ elk/__main__.py
@@ -1,14 +1,10 @@
-from .extraction.extraction_main import run as run_extraction
-from .extraction.parser import get_extraction_parser
-from .training.parser import get_training_parser
-from .training.train import train
+from .argparsers import get_extraction_parser, get_training_parser
+from .files import args_to_uuid
+from .list import list_runs
 from argparse import ArgumentParser
 from contextlib import nullcontext, redirect_stdout
-from elk.files import args_to_uuid
-from transformers import AutoConfig, PretrainedConfig
 import logging
 import os
-import torch.distributed as dist
 
 
 def run():
@@ -41,7 +37,15 @@ def run():
     subparsers.add_parser(
         "eval", help="Evaluate a set of ELK probes generated by `elk train`."
     )
+    subparsers.add_parser("list", help="List all cached runs.")
     args = parser.parse_args()
+
+    # `elk list` is a special case
+    if args.command == "list":
+        list_runs(args)
+        return
+
+    from transformers import AutoConfig, PretrainedConfig
 
     if model := getattr(args, "model", None):
         config = AutoConfig.from_pretrained(model)
@@ -58,7 +62,10 @@ def run():
             args.layers = list(range(0, num_layers, args.layer_stride))
 
     # Support both distributed and non-distributed training
+    import torch.distributed as dist
+
     local_rank = os.environ.get("LOCAL_RANK")
+
     if local_rank is not None:
         dist.init_process_group("nccl")
         local_rank = int(local_rank)
@@ -74,13 +81,17 @@ def run():
 
     # Prevent printing from processes other than the first one
     with redirect_stdout(None) if local_rank != 0 else nullcontext():
+        # Print all arguments
         for key in list(vars(args).keys()):
             print("{}: {}".format(key, vars(args)[key]))
 
         if local_rank != 0:
             logging.getLogger("transformers").setLevel(logging.ERROR)
 
-        # TODO: Implement the rest of the CLI
+        # Import here and not at the top to speed up `elk list`
+        from .extraction.extraction_main import run as run_extraction
+        from .training.train import train
+
         if args.command == "extract":
             run_extraction(args)
         elif args.command == "train":
@@ -99,6 +110,7 @@ def run():
                 train(args)
 
         elif args.command == "eval":
+            # TODO: Implement evaluation script
             raise NotImplementedError
         else:
             raise ValueError(f"Unknown command {args.command}")
